requirements.txt
# Core
langchain==0.3.25
langchain-community==0.3.24
langchain-huggingface==0.1.2

# Document processing
pypdf==5.1.0
python-docx==1.1.2
unstructured==0.16.11

# Embeddings
sentence-transformers==3.3.1

# Vector database
chromadb==0.5.23

# LLM providers (pick what you need)
langchain-openai==0.3.0
huggingface-hub==0.27.1

# API
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.20

# Frontend
streamlit==1.41.1

# Evaluation
ragas==0.2.10
datasets==3.2.0

# Keyword search
rank-bm25==0.2.2

# Utilities
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.1
